Change variable type of ram to long

+++ api/api-default-impl/src/main/java/net/suqatri/cloud/api/impl/service/CloudService.java
@@ -28,7 +28,7 @@ public class CloudService extends RBucketObject implements ICloudService {
     private int maxPlayers = -1;
     private ServiceState serviceState = ServiceState.UNKNOWN;
     private int onlineCount = -1;
-    private int maxRam = -1;
+    private long maxRam = -1;
     private long ramUsage = -1;
     private UUID nodeId;
     private Collection<UUID> consoleNodeListenerIds = new ArrayList<>();
